requirements.txt
# Django
Django==3.0.5
django-choices==1.7.1

# Django jet
django-3-jet==1.0.8

# utils
django-extensions==2.2.9
django-debug-toolbar==2.2
django-simple-history==2.8.0

# Django rest framework
djangorestframework==3.11.0
django-filter==2.2.0
typing==3.7.4.1
